Removing unnecessary functions
